backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dagster>=1.11.6",
    "dagster-dg-cli>=1.11.6",
    "dagster-duckdb>=0.27.6",
    "dagster-webserver>=1.11.6",
    "fastparquet>=2024.11.0",
    "ipykernel>=6.30.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "peft>=0.17.0",
    "pymongo>=4.14.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "transformers>=4.55.2",
    "trl>=0.21.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]
ner = [
    "spacy>=3.8.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "backend"
registry_modules = [
    "backend.components.*",
]
